requirements.txt
praw==7.7.1
Pillow==10.2.0
python-dotenv==1.0.0
schedule==1.2.0
pytz==2023.4